backend/requirements.txt
django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
python-decouple==3.8
httpx==0.27.0
beautifulsoup4==4.12.3
sqlparse==0.5.5
asgiref==3.11.1
sniffio==1.3.1
certifi==2026.2.25
idna==3.11
anyio==4.13.0
httpcore==1.0.9
h11==0.16.0
soupsieve==2.8.3
typing_extensions==4.15.0